clean vs warning C4239;  user's operator only care about lvalues without temporary;

+++ src/include/fn_parse.h
@@ -378,7 +378,7 @@ namespace FNLog
         char buff[50];
         memcpy(buff, port_begin, end - port_begin);
         buff[end - port_begin] = '\0';
-        port = htons(atoi(buff));
+        port = htons((unsigned short)atoi(buff));
         ip = result_ip.first;
         return;
     }
